// Hippocamp.Web/src/app/shared/generated/model/project-document-dto.ts
/**
 * Hippocamp.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProjectDto } from '././project-dto';
import { FileResourceDto } from '././file-resource-dto';

export class ProjectDocumentDto { 
    ProjectDocumentID?: number;
    FileResource?: FileResourceDto;
    Project?: ProjectDto;
    DisplayName?: string;
    UploadDate?: string;
    DocumentDescription?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}